Split up difference.h into difference.h and change.h
